refactor: Refine dictionary column.
